merge from master

+++ src/view/GameView.java
@@ -27,7 +27,7 @@ import model.objects.SnakeTail;
 
 import util.GameEvent;
 import util.Vector2D;
-import view.Figures.*;
+import view.figures.*;
 
 /**
  * This is a class that just contains the game itself
@@ -191,7 +191,7 @@ implements MouseWheelListener, MouseMotionListener, MouseListener, GameObserver,
 			addItem(thing);
 		}
 		System.out.println("Addworld i GameView");
-		worldSize=world.getWorldSize();
+		worldSize=(int)world.getWorldSize();
 	}
